refactor: split EditingPanel into separate components
